# web/packages/api/openapi.yaml
openapi: 3.1.0
info:
  title: Teggo B2B Commerce API
  version: 0.1.0
  description: >
    The single source of truth for the Teggo API contract. The TypeScript
    client used by both frontends is generated from this file (do not hand-edit
    the generated schema). Grown module-by-module as the backend lands.
servers:
  - url: http://localhost:8080
    description: Local Go API
...
components:
  securitySchemes:
    adminAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    storefrontAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    vendorAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    OrderAddressInput:
      type: object
      required: [line1, city, country]
      properties:
        line1: { type: string }
        line2: { type: string, nullable: true }
        city: { type: string }
        region: { type: string, nullable: true }
        postal_code: { type: string, nullable: true }
        country: { type: string, minLength: 2, maxLength: 2 }
    PlaceOrderRequest:
      type: object
      properties:
        po_number: { type: string, nullable: true }
        requested_delivery_date: { type: string, format: date-time, nullable: true }
        billing_address: { $ref: '#/components/schemas/OrderAddressInput' }
        shipping_address: { $ref: '#/components/schemas/OrderAddressInput' }
        cost_center: { type: string, nullable: true }
        display_currency: { type: string, nullable: true, description: 'Currency the buyer was quoted in; locks the FX rate onto the order.' }
...
    RfqItem:
      type: object
      required: [id, product_id, sku, name, quantity, unit]
      properties:
        id: { type: integer, format: int64 }
        product_id: { type: integer, format: int64 }
        sku: { type: string }
        name: { type: string }
        quantity: { type: string }
        unit: { type: string }
        target_price: { type: string, nullable: true }
        notes: { type: string, nullable: true }
    RfqDetail:
      type: object
      required: [public_id, status, items]
      properties:
        public_id: { type: string, format: uuid }
        status: { type: string, enum: [draft, submitted, quoted, accepted, declined, expired, cancelled] }
        notes: { type: string, nullable: true }
        items: { type: array, items: { $ref: '#/components/schemas/RfqItem' } }
    RfqSummary:
      type: object
      required: [id, public_id, status]
      properties:
        id: { type: integer, format: int64 }
        public_id: { type: string, format: uuid }
        customer_id: { type: integer, format: int64 }
        status: { type: string }
        notes: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    CreateRFQ:
      type: object
      required: [items]
      properties:
        notes: { type: string }
        items:
          type: array
          items:
            type: object
            required: [product_public_id, quantity]
            properties:
              product_public_id: { type: string, format: uuid }
              quantity: { type: string }
              target_price: { type: string, nullable: true }
              notes: { type: string, nullable: true }
    QuoteItem:
      type: object
      required: [id, product_id, sku, name, quantity, unit, unit_price, discount, row_total]
      properties:
        id: { type: integer, format: int64 }
        product_id: { type: integer, format: int64 }
        sku: { type: string }
        name: { type: string }
        quantity: { type: string }
        unit: { type: string }
        unit_price: { type: string }
        discount: { type: string }
        row_total: { type: string }
    QuoteDetail:
      type: object
      required: [id, public_id, status, currency, version, subtotal, items]
      properties:
        id: { type: integer, format: int64 }
        public_id: { type: string, format: uuid }
        status: { type: string, enum: [draft, sent, revised, accepted, declined, expired] }
        currency: { type: string }
        version: { type: integer }
        subtotal: { type: string }
        valid_until: { type: string, format: date-time, nullable: true }
        items: { type: array, items: { $ref: '#/components/schemas/QuoteItem' } }
    QuoteSummary:
      type: object
      required: [id, public_id, status, currency, version, subtotal]
      properties:
        id: { type: integer, format: int64 }
        public_id: { type: string, format: uuid }
        customer_id: { type: integer, format: int64 }
        rfq_id: { type: integer, format: int64, nullable: true }
        status: { type: string }
        currency: { type: string }
        version: { type: integer }
        subtotal: { type: string }
        valid_until: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
    QuoteLineInput:
      type: object
      required: [product_id, unit_price]
      properties:
        product_id: { type: integer, format: int64 }
        quantity: { type: string }
        unit: { type: string }
        unit_price: { type: string }
        discount: { type: string }
    CreateQuote:
      type: object
      required: [customer_id, items]
      properties:
        customer_id: { type: integer, format: int64 }
        currency: { type: string }
        valid_until: { type: string, format: date-time, nullable: true }
        items: { type: array, items: { $ref: '#/components/schemas/QuoteLineInput' } }
    EditQuote:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/QuoteLineInput' } }
    SendQuote:
      type: object
      properties:
        valid_until: { type: string, format: date-time, nullable: true }
    OrderItem:
      type: object
      required: [id, product_id, sku, name, quantity, unit, unit_price, row_total]
      properties:
        id: { type: integer, format: int64 }
        product_id: { type: integer, format: int64 }
        sku: { type: string }
        name: { type: string }
        quantity: { type: string }
        unit: { type: string }
        unit_price: { type: string }
        tax_amount: { type: string }
        row_total: { type: string }
    OrderDetail:
      type: object
      required: [id, public_id, status, currency, subtotal, grand_total, items]
      properties:
        id: { type: integer, format: int64 }
        public_id: { type: string, format: uuid }
        status: { type: string, enum: [pending, confirmed, processing, shipped, delivered, closed, on_hold, cancelled] }
        currency: { type: string }
        subtotal: { type: string }
        tax_total: { type: string }
        shipping_total: { type: string }
        grand_total: { type: string }
        quote_id: { type: integer, format: int64, nullable: true }
        paid: { type: boolean }
        invoice_public_id: { type: string, format: uuid, nullable: true }
        items: { type: array, items: { $ref: '#/components/schemas/OrderItem' } }
    OrderSummary:
      type: object
      required: [id, public_id, status, currency, grand_total]
      properties:
        id: { type: integer, format: int64 }
        public_id: { type: string, format: uuid }
        status: { type: string }
        currency: { type: string }
        subtotal: { type: string }
        grand_total: { type: string }
        discount_total: { type: string }
        promotion_code: { type: string, nullable: true }
        placed_by_sales_rep_id: { type: integer, format: int64, nullable: true, description: 'Set when a sales rep placed the order on the customer’s behalf.' }
        quote_id: { type: integer, format: int64, nullable: true }
        created_at: { type: string, format: date-time }
    OrderStatusPatch:
      type: object
      required: [status]
      properties:
        status: { type: string }
        note: { type: string, nullable: true }
    CreateOrderOnBehalf:
      type: object
      required: [customer_id, items]
      properties:
        customer_id: { type: integer, format: int64 }
        currency: { type: string }
        po_number: { type: string, nullable: true }
        requested_delivery_date: { type: string, format: date, nullable: true }
        items:
          type: array
          items:
            type: object
            required: [product_id, unit_price]
            properties:
              product_id: { type: integer, format: int64 }
              quantity: { type: string }
              unit: { type: string }
              unit_price: { type: string }
    ListWrapperRfqSummary:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/RfqSummary' } }
    ListWrapperQuoteSummary:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/QuoteSummary' } }
    ListWrapperOrderSummary:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/OrderSummary' } }
...
  responses:
    ErrorResponse:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
paths:
...
  /admin/rfqs:
    get:
      tags: [Sales]
      operationId: adminListRfqs
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ListWrapperRfqSummary' } } } }
  /admin/rfqs/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [Sales]
      operationId: adminGetRfq
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/RfqDetail' } } } }
        '404': { $ref: '#/components/responses/ErrorResponse' }
  /admin/rfqs/{id}/quote:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    post:
      tags: [Sales]
      operationId: adminQuoteFromRfq
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/QuoteDetail' } } } }
  /admin/quotes:
    get:
      tags: [Sales]
      operationId: adminListQuotes
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ListWrapperQuoteSummary' } } } }
    post:
      tags: [Sales]
      operationId: adminCreateQuote
      security: [{ adminAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateQuote' } } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/QuoteDetail' } } } }
  /admin/quotes/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [Sales]
      operationId: adminGetQuote
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/QuoteDetail' } } } }
        '404': { $ref: '#/components/responses/ErrorResponse' }
    put:
      tags: [Sales]
      operationId: adminEditQuote
      security: [{ adminAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/EditQuote' } } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/QuoteDetail' } } } }
        '409': { $ref: '#/components/responses/ErrorResponse' }
  /admin/quotes/{id}/send:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    post:
      tags: [Sales]
      operationId: adminSendQuote
      security: [{ adminAuth: [] }]
      requestBody: { required: false, content: { application/json: { schema: { $ref: '#/components/schemas/SendQuote' } } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/QuoteDetail' } } } }
        '409': { $ref: '#/components/responses/ErrorResponse' }
  /admin/orders:
    get:
      tags: [Sales]
      operationId: adminListOrders
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ListWrapperOrderSummary' } } } }
    post:
      tags: [Sales]
      operationId: adminCreateOrder
      security: [{ adminAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateOrderOnBehalf' } } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OrderDetail' } } } }
  /admin/orders/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [Sales]
      operationId: adminGetOrder
      security: [{ adminAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OrderDetail' } } } }
        '404': { $ref: '#/components/responses/ErrorResponse' }
  /admin/orders/{id}/status:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
    patch:
      tags: [Sales]
      operationId: adminPatchOrderStatus
      security: [{ adminAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/OrderStatusPatch' } } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OrderDetail' } } } }
        '409': { $ref: '#/components/responses/ErrorResponse' }
  /storefront/rfqs:
    get:
      tags: [Sales]
      operationId: storefrontListRfqs
      security: [{ storefrontAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ListWrapperRfqSummary' } } } }
    post:
      tags: [Sales]
      operationId: storefrontCreateRfq
      security: [{ storefrontAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateRFQ' } } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/RfqDetail' } } } }
  /storefront/rfqs/{publicID}:
    parameters:
      - { name: publicID, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Sales]
      operationId: storefrontGetRfq
      security: [{ storefrontAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/RfqDetail' } } } }
        '404': { $ref: '#/components/responses/ErrorResponse' }
  /storefront/rfqs/{publicID}/submit:
    parameters:
      - { name: publicID, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [Sales]
      operationId: storefrontSubmitRfq
      security: [{ storefrontAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/RfqDetail' } } } }
        '409': { $ref: '#/components/responses/ErrorResponse' }
  /storefront/quotes/{publicID}:
    parameters:
      - { name: publicID, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Sales]
      operationId: storefrontGetQuote
      security: [{ storefrontAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/QuoteDetail' } } } }
        '404': { $ref: '#/components/responses/ErrorResponse' }
  /storefront/quotes/{publicID}/accept:
    parameters:
      - { name: publicID, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [Sales]
      operationId: storefrontAcceptQuote
      security: [{ storefrontAuth: [] }]
      responses:
        '200': { description: Order created, content: { application/json: { schema: { $ref: '#/components/schemas/OrderDetail' } } } }
        '409': { $ref: '#/components/responses/ErrorResponse' }
  /storefront/quotes/{publicID}/decline:
    parameters:
      - { name: publicID, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [Sales]
      operationId: storefrontDeclineQuote
      security: [{ storefrontAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/QuoteDetail' } } } }
        '409': { $ref: '#/components/responses/ErrorResponse' }
  /storefront/orders:
    get:
      tags: [Sales]
      operationId: storefrontListOrders
      security: [{ storefrontAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ListWrapperOrderSummary' } } } }
    post:
      tags: [Sales]
      operationId: storefrontPlaceOrder
      summary: Place an order from the active cart
      description: Converts the customer's active cart into a pending order in one transaction. Addresses default to the customer's defaults when omitted.
      security: [{ storefrontAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlaceOrderRequest' }
      responses:
        '200': { description: Order created, content: { application/json: { schema: { $ref: '#/components/schemas/OrderDetail' } } } }
        '422': { $ref: '#/components/responses/ErrorResponse' }
  /storefront/orders/{publicID}:
    parameters:
      - { name: publicID, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Sales]
      operationId: storefrontGetOrder
      security: [{ storefrontAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OrderDetail' } } } }
        '404': { $ref: '#/components/responses/ErrorResponse' }
...
  /storefront/quotes:
    get:
      tags: [Sales]
      operationId: storefrontListQuotes
      security: [{ storefrontAuth: [] }]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/ListWrapperQuoteSummary' } } } }
...